spec4_backend/pyproject.toml
[tool.poetry]
name = "spec4-backend"
version = "0.1.0"
description = "FastAPI backend for RAG agent integration"
authors = ["Physical-AI-Textbook Team"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
anthropic = "^0.21.3"
python-dotenv = "^1.0.0"
openai = "^1.0.0"
pydantic = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"